Moved generic code to new file in the contexts folder and fixed imports
