+++ toktagger/ui/src/app/contexts/NavAdapterContext.tsx
@@ -1,0 +1,41 @@
+"use client";
+
+import React, { createContext, useContext } from "react";
+import { useSample } from "@/app/contexts/SampleContext";
+import { type NavAdapter } from "@/types";
+
+const NavAdapterContext = createContext<NavAdapter | null>(null);
+
+export function NavAdapterProvider({
+  value,
+  children,
+}: {
+  value: NavAdapter;
+  children: React.ReactNode;
+}) {
+  return (
+    <NavAdapterContext.Provider value={value}>
+      {children}
+    </NavAdapterContext.Provider>
+  );
+}
+
+export function useNavAdapterOptional(): NavAdapter | null {
+  return useContext(NavAdapterContext);
+}
+
+export function useNavAdapter(): NavAdapter {
+  const navAdapter = useNavAdapterOptional();
+  const { annotations, setAnnotations } = useSample();
+
+  if (navAdapter) {
+    return navAdapter;
+  }
+
+  return {
+    getAnnotations: () => annotations,
+    clear: () => {
+      setAnnotations(() => []);
+    },
+  };
+}
