+++ toktagger/ui/src/app/video/components/video-nav-adapter.tsx
@@ -1,49 +1,10 @@
 "use client";
 
-import React, { createContext, useContext } from "react";
+import React from "react";
 import { useSample } from "@/app/contexts/SampleContext";
-import {
-  type Annotation,
-  type NavAdapter,
-  VideoBoundingBoxSchema,
-} from "@/types";
+import { NavAdapterProvider } from "@/app/contexts/NavAdapterContext";
+import { type Annotation, type NavAdapter, VideoBoundingBoxSchema } from "@/types";
 import { useVideoSession } from "./video-session";
-
-const NavAdapterContext = createContext<NavAdapter | null>(null);
-
-export function NavAdapterProvider({
-  value,
-  children,
-}: {
-  value: NavAdapter;
-  children: React.ReactNode;
-}) {
-  return (
-    <NavAdapterContext.Provider value={value}>
-      {children}
-    </NavAdapterContext.Provider>
-  );
-}
-
-export function useNavAdapterOptional(): NavAdapter | null {
-  return useContext(NavAdapterContext);
-}
-
-export function useNavAdapter(): NavAdapter {
-  const navAdapter = useNavAdapterOptional();
-  const { annotations, setAnnotations } = useSample();
-
-  if (navAdapter) {
-    return navAdapter;
-  }
-
-  return {
-    getAnnotations: () => annotations,
-    clear: () => {
-      setAnnotations(() => []);
-    },
-  };
-}
 
 export function VideoNavAdapterBridge({
   children,
